refactor: move damage particle effects to (separate) authority system
